properly splitting json files
also decided to pull documents from the source instead of having them in
this repo, so removed those documents and changed the file path to get
to the CVE repo.

There are way more documents now from splitting, so it is taking some
time to get them all in the database.